+++ src-docs/src/views/image/image.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import {
+  EuiImage,
+} from '../../../../src/components';
+
+export default () => (
+  <EuiImage
+    size="l"
+    hasShadow
+    caption="I am a starship. Zooom!"
+    alt="Accessible image alt goes here"
+    url="https://i.imgur.com/4qx7HhE.jpg"
+  />
+);